Unify plotTimeSeriesWithObs/plotTimeSeriesDebug into plotTimeSeries with backward-compat aliases
