+++ src/Visualization/Visualization.jl
@@ -12,10 +12,11 @@ plotting model outputs, diagnostics, and experiment metadata via `Plots.jl`.
 
 # Dependencies
 ## Internal (within `Sindbad`)
-- `namedTupleToFlareJSON` and `getAllVariables` (`plotFromSindbadInfo.jl`) have no external plotting dependency and are always available.
+- `namedTupleToFlareJSON` and `getAllVariables` (`utilsVisualization.jl`) have no external plotting dependency and are always available.
 
 ## Optional (via `SindbadPlotsExt`)
-- `plotPerformanceHistograms`, `plotTimeSeriesWithObs`, `plotTimeSeriesDebug`, and `plotIOModelStructure` dispatch on a `VisualizationTypes` backend value read from `info.helpers.run.visualization_backend` (set from `exe_rules.visualization_backend` in the experiment config, defaulting to `"Types"`). The default `VisualizationTypes` backend logs an info message and returns `nothing`. Setting `visualization_backend = "Plots"` in the config selects the `VisualizationPlots` backend, whose real plotting methods are only added once `Plots.jl` is loaded, at which point Julia's package extension mechanism loads `SindbadPlotsExt` (see root `Project.toml` `[weakdeps]` + `[extensions]`). `using Sindbad` alone is sufficient once `Plots` is also loaded in the session — end users typically should not `using SindbadPlotsExt` directly.
+- `plotPerformanceHistograms`, `plotTimeSeries`, and `plotIOModelStructure` dispatch on a `VisualizationTypes` backend value read from `info.helpers.run.visualization_backend` (set from `exe_rules.visualization_backend` in the experiment config, defaulting to `"Types"`). The default `VisualizationTypes` backend logs an info message and returns `nothing`. Setting `visualization_backend = "Plots"` in the config selects the `VisualizationPlots` backend, whose real plotting methods are only added once `Plots.jl` is loaded, at which point Julia's package extension mechanism loads `SindbadPlotsExt` (see root `Project.toml` `[weakdeps]` + `[extensions]`). `using Sindbad` alone is sufficient once `Plots` is also loaded in the session — end users typically should not `using SindbadPlotsExt` directly.
+- `plotTimeSeriesWithObs` and `plotTimeSeriesDebug` are deprecated-but-supported aliases of `plotTimeSeries`, kept for backward compatibility with their original signatures.
 
 # Notes
 - Additional plot recipes are being added progressively; current focus is
@@ -36,6 +37,7 @@ module Visualization
     using ..Types
 
     export plotPerformanceHistograms
+    export plotTimeSeries
     export plotTimeSeriesWithObs
     export plotTimeSeriesDebug
     export plotIOModelStructure
@@ -62,19 +64,35 @@ module Visualization
         return plotPerformanceHistograms(out_opti, _resolvedBackend(plotPerformanceHistograms, backend, out_opti))
     end
 
-    function plotTimeSeriesWithObs(out_opti)
+    function plotTimeSeries(out_opti)
         backend = out_opti.info.helpers.run.visualization_backend
-        return plotTimeSeriesWithObs(out_opti, _resolvedBackend(plotTimeSeriesWithObs, backend, out_opti))
+        return plotTimeSeries(out_opti, _resolvedBackend(plotTimeSeries, backend, out_opti))
     end
+
+    function plotTimeSeries(out, cost_options)
+        backend = out.info.helpers.run.visualization_backend
+        return plotTimeSeries(out, cost_options, _resolvedBackend(plotTimeSeries, backend, out, cost_options))
+    end
+
+    function plotTimeSeries(info, opt_dat, def_dat)
+        backend = info.helpers.run.visualization_backend
+        return plotTimeSeries(info, opt_dat, def_dat, _resolvedBackend(plotTimeSeries, backend, info, opt_dat, def_dat))
+    end
+
+    # Backward-compatible aliases: old names/signatures forward onto the unified
+    # plotTimeSeries dispatch family. plotTimeSeriesWithObs(out_opti) and
+    # plotTimeSeriesDebug(info, opt_dat, def_dat) match new entry points exactly, so they
+    # inherit backend resolution/fallback for free. The legacy 3-arg
+    # plotTimeSeriesWithObs(out, obs_array, cost_options) doesn't match any new entry
+    # point's shape (the 2-arg plotTimeSeries derives obs_array from
+    # out.observation instead), so it resolves its own backend and forwards to the
+    # explicit-argument method directly, preserving the exact old behavior/signature.
+    plotTimeSeriesWithObs(out_opti) = plotTimeSeries(out_opti)
+    plotTimeSeriesDebug(info, opt_dat, def_dat) = plotTimeSeries(info, opt_dat, def_dat)
 
     function plotTimeSeriesWithObs(out, obs_array, cost_options)
         backend = out.info.helpers.run.visualization_backend
-        return plotTimeSeriesWithObs(out, obs_array, cost_options, _resolvedBackend(plotTimeSeriesWithObs, backend, out, obs_array, cost_options))
-    end
-
-    function plotTimeSeriesDebug(info, opt_dat, def_dat)
-        backend = info.helpers.run.visualization_backend
-        return plotTimeSeriesDebug(info, opt_dat, def_dat, _resolvedBackend(plotTimeSeriesDebug, backend, info, opt_dat, def_dat))
+        return plotTimeSeries(out.info, obs_array, cost_options, out.output, _resolvedBackend(plotTimeSeries, backend, out.info, obs_array, cost_options, out.output))
     end
 
     plotIOModelStructure(info) = plotIOModelStructure(info, :compute)
@@ -86,31 +104,54 @@ module Visualization
 
     # Default (no-op) backend: logs an info message instead of plotting. Used whenever
     # `visualization_backend` isn't configured, or a configured backend's extension isn't loaded.
+    #
+    # `configured_backend` is re-read from `info`/`out`/`out_opti` (not the dispatch arg, which
+    # `_resolvedBackend` may have already substituted down to `VisualizationTypes()`), so the
+    # message can distinguish "nothing was ever configured" from "a specific backend was
+    # configured but its extension isn't loaded".
+    function _visualizationFallbackMessage(fn, configured_backend)
+        if configured_backend isa VisualizationTypes
+            return "No visualization backend configured; skipping `$(nameof(fn))`. Set `visualization_backend` in the experiment config (e.g. `\"Plots\"`) and `using Plots` in your session to enable this plot."
+        else
+            backend_name = String(nameof(typeof(configured_backend)))
+            pkg_name = startswith(backend_name, "Visualization") ? backend_name[(length("Visualization") + 1):end] : backend_name
+            return "visualization_backend is set to `$(backend_name)`, but `$(pkg_name)` is not loaded; skipping `$(nameof(fn))`. Run `using $(pkg_name)` in your session to enable this plot."
+        end
+    end
+
     function plotPerformanceHistograms(out_opti, ::VisualizationTypes)
-        print_info(plotPerformanceHistograms, @__FILE__, @__LINE__, "No visualization backend loaded; skipping `plotPerformanceHistograms`. Set `visualization_backend` in the experiment config (e.g. `\"Plots\"`) and `using Plots` in your session to enable this plot.", n_f=4)
+        print_info(plotPerformanceHistograms, @__FILE__, @__LINE__, _visualizationFallbackMessage(plotPerformanceHistograms, out_opti.info.helpers.run.visualization_backend), n_f=4)
         return nothing
     end
 
-    function plotTimeSeriesWithObs(out_opti, ::VisualizationTypes)
-        print_info(plotTimeSeriesWithObs, @__FILE__, @__LINE__, "No visualization backend loaded; skipping `plotTimeSeriesWithObs`. Set `visualization_backend` in the experiment config (e.g. `\"Plots\"`) and `using Plots` in your session to enable this plot.", n_f=4)
+    function plotTimeSeries(out_opti, ::VisualizationTypes)
+        print_info(plotTimeSeries, @__FILE__, @__LINE__, _visualizationFallbackMessage(plotTimeSeries, out_opti.info.helpers.run.visualization_backend), n_f=4)
         return nothing
     end
 
-    function plotTimeSeriesWithObs(out, obs_array, cost_options, ::VisualizationTypes)
-        print_info(plotTimeSeriesWithObs, @__FILE__, @__LINE__, "No visualization backend loaded; skipping `plotTimeSeriesWithObs`. Set `visualization_backend` in the experiment config (e.g. `\"Plots\"`) and `using Plots` in your session to enable this plot.", n_f=4)
+    function plotTimeSeries(out, cost_options, ::VisualizationTypes)
+        print_info(plotTimeSeries, @__FILE__, @__LINE__, _visualizationFallbackMessage(plotTimeSeries, out.info.helpers.run.visualization_backend), n_f=4)
         return nothing
     end
 
-    function plotTimeSeriesDebug(info, opt_dat, def_dat, ::VisualizationTypes)
-        print_info(plotTimeSeriesDebug, @__FILE__, @__LINE__, "No visualization backend loaded; skipping `plotTimeSeriesDebug`. Set `visualization_backend` in the experiment config (e.g. `\"Plots\"`) and `using Plots` in your session to enable this plot.", n_f=4)
+    function plotTimeSeries(info, opt_dat, def_dat, ::VisualizationTypes)
+        print_info(plotTimeSeries, @__FILE__, @__LINE__, _visualizationFallbackMessage(plotTimeSeries, info.helpers.run.visualization_backend), n_f=4)
+        return nothing
+    end
+
+    # Needed so the legacy 3-arg plotTimeSeriesWithObs(out, obs_array, cost_options)
+    # alias (which forwards straight to this arity, bypassing the arity-2 wrapper above)
+    # still degrades gracefully when no Plots backend is loaded.
+    function plotTimeSeries(info, obs_array, cost_options, def_dat, ::VisualizationTypes)
+        print_info(plotTimeSeries, @__FILE__, @__LINE__, _visualizationFallbackMessage(plotTimeSeries, info.helpers.run.visualization_backend), n_f=4)
         return nothing
     end
 
     function plotIOModelStructure(info, which_function, which_field, ::VisualizationTypes)
-        print_info(plotIOModelStructure, @__FILE__, @__LINE__, "No visualization backend loaded; skipping `plotIOModelStructure`. Set `visualization_backend` in the experiment config (e.g. `\"Plots\"`) and `using Plots` in your session to enable this plot.", n_f=4)
+        print_info(plotIOModelStructure, @__FILE__, @__LINE__, _visualizationFallbackMessage(plotIOModelStructure, info.helpers.run.visualization_backend), n_f=4)
         return nothing
     end
 
-    include("plotFromSindbadInfo.jl")
+    include("utilsVisualization.jl")
 
 end # module Visualization
